Use OpenAI responses for web search
